Extract PrivateMoodTaxonomyService from SpotifyTopPlaylistsService
SpotifyTopPlaylistsService had grown to ~1,600 lines by absorbing the
entire Private Mood Taxonomy pipeline (signal building, per-mood candidate
selection, lyric analysis and reranking, frontier exploration, and its
large keyword-weight tables) alongside the unrelated yearly, top, and
forgotten-obsessions playlist flows.

Move the Private Mood Taxonomy subsystem into a dedicated
PrivateMoodTaxonomyService, leaving SpotifyTopPlaylistsService focused on
the top/yearly/forgotten flows (now ~440 lines). Shared helpers
(normalizeToken/normalizedKey/toSong, SECONDS_PER_DAY, and playlist
get-or-create with per-name creation locking) are extracted into
PlaylistSupport so neither service duplicates them.

- JobService now injects PrivateMoodTaxonomyService for the private-mood
  job; all other call sites are unchanged.
- LyricsService references the profile type at its new location.
- The unused LyricsService dependency is dropped from
  SpotifyTopPlaylistsService's constructor.
- Private-mood tests move to PrivateMoodTaxonomyServiceTest; JobServiceTest
  and JobsControllerIT wire the new service. Behavior is unchanged;
  build, tests, ktfmt, and the 90% coverage gate all pass.
